autoscaler-node/auto_scaler_project/requirements.txt
alembic==1.14.0
amqp==5.2.0
aniso8601==9.0.1
attrs==24.2.0
billiard==4.2.1
blinker==1.4
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
circuitbreaker==2.0.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.0
cryptography==42.0.8
cycler==0.12.1
dateparser==1.2.0
dnspython==2.7.0
email-validator==1.1.3
Flask==2.0.2
Flask-Dance==5.1.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Minify==0.37
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fonttools==4.53.1
greenlet==3.1.1
gunicorn==20.1.0
htmlmin==0.1.12
httmock==1.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.0.3
jsmin==3.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kombu==5.4.2
lesscpy==0.15.1
Mako==1.3.6
MarkupSafe==3.0.2
matplotlib==3.9.2
numpy==2.1.1
oauthlib==3.2.2
oci==2.135.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
ply==3.11
prometheus-api-client==0.5.5
prompt_toolkit==3.0.48
pycparser==2.22
pyOpenSSL==24.2.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2024.2
PyYAML==6.0.2
rcssmin==1.2.0
redis==5.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.21.0
six==1.16.0
SQLAlchemy==1.4.29
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.3
URLObject==2.4.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.0.3
WTForms==3.0.1
xxhash==3.5.0